[#7] Move getChartData to utilities
