+++ client/src/components/charts/DashLineChart.jsx
@@ -1,32 +1,10 @@
 import React, { Component, PropTypes } from 'react';
 import { LineChart } from 'react-d3';
-
-const getChartData = (visualisation, datasets) => {
-  const datasetID = visualisation.sourceDatasetX;
-  const dataset = datasets[datasetID];
-  const columnIndex = visualisation.datasetColumnX;
-  const data = dataset.columns[columnIndex].values;
-  const dataValues = [];
-  const output = [];
-
-  data.map((entry, index) => {
-    dataValues.push({
-      x: index,
-      y: parseInt(entry, 10),
-    });
-  });
-
-  output.push({
-    name: 'series1',
-    values: dataValues,
-  });
-
-  return output;
-};
+import * as chart from '../../utilities/chart';
 
 export default class DashLineChart extends Component {
   render() {
-    const chartData = getChartData(this.props.visualisation, this.props.datasets);
+    const chartData = chart.getChartData(this.props.visualisation, this.props.datasets);
     const visualisation = this.props.visualisation;
     const gridHorizontal = true;
     const gridVertical = true;
